Expand paths before adding them to the goto file
For example, if you'd say "goto-tool add foobar ." before, it would actually
create the listing entry "foobar: .", which is rather useless, imho.
Now, the path is expanded with "os.path.abspath()", which makes the tool
more usable. If for some reason (because I can't think of one) it would
be desirable to add a relative path to the listing, a command line flag
should be added to force that behavior.