Support symlinks as target if wanted
In our daily live we may have situations where we have different
versions of folders and one central folder which is a symlink and
points to the current version of our thingy.

To support such scenarios the --soft flag is added to the add command.
If --soft is supplied, goto-tool does not make an attempt to resolve
the real path of the target. Additionally, if '.' is given as target
directory, goto-tool tries to resolve the current symlink name by
getting PWD from the shell. This is not perfect, because it depends
on the shell but it's a nice feature for those shells that are not
resolving symlinks in PWD.

If a shell does resolve symlinks in PWD, one can still add the symlink
from the target directory which will work regardless of the shell.